Sample GRC knowledge graph showing automated test infrastructure
This knowledge graph attempts to show the tests associated with GRC
repositories and the integration with e2e testing and fast forwarding to
release branches.  This is a work in progress.  The script in the bin
directory is the beginning of a way to automate creation of the xml file
but many updates to the kg were manually made after running the script.

Signed-off-by: Gus Parvin <gparvin@redhat.com>